+++ client/v2/client_test.go
@@ -13,7 +13,7 @@ import (
 	"go.sia.tech/core/types"
 	"go.sia.tech/indexd/client/v2"
 	"go.sia.tech/indexd/hosts"
-	"go.sia.tech/indexd/internal/testutils"
+	"go.sia.tech/indexd/testutils"
 	"go.uber.org/zap/zaptest"
 	"lukechampine.com/frand"
 )
