+++ contracts/maintenance_test.go
@@ -12,8 +12,8 @@ import (
 	"go.sia.tech/indexd/alerts"
 	"go.sia.tech/indexd/contracts"
 	"go.sia.tech/indexd/hosts"
-	"go.sia.tech/indexd/internal/testutils"
 	"go.sia.tech/indexd/subscriber"
+	"go.sia.tech/indexd/testutils"
 	"go.uber.org/zap/zaptest"
 )
