+++ hosts/manager_test.go
@@ -16,9 +16,9 @@ import (
 	"go.sia.tech/indexd/alerts"
 	"go.sia.tech/indexd/contracts"
 	"go.sia.tech/indexd/hosts"
-	"go.sia.tech/indexd/internal/testutils"
-	"go.sia.tech/indexd/internal/testutils/mock"
 	"go.sia.tech/indexd/subscriber"
+	"go.sia.tech/indexd/testutils"
+	"go.sia.tech/indexd/testutils/mock"
 	"go.uber.org/zap/zaptest"
 )
